closing file always removes correct CSVDefinition Previously RemoveCsvDef didn't handle the edge case where the buffer closed was not the currently open buffer, and removed the CSVDefinition associated with the current buffer even if another buffer was closed instead.
Changed the GetFilePath method of NotepadPPGateway handle IntPtr
instead of int, allowing the new feature to work in both 32-bit
and 64-bit Notepad++.

+++ CSVLintNppPlugin/PluginInfrastructure/NotepadPPGateway.cs
@@ -10,7 +10,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
 		void FileNew();
 
 		string GetCurrentFilePath();
-		unsafe string GetFilePath(int bufferId);
+		unsafe string GetFilePath(IntPtr bufferId);
 		void SetCurrentLanguage(LangType language);
 	}
 
@@ -40,7 +40,7 @@ namespace Kbg.NppPluginNET.PluginInfrastructure
 		/// <summary>
 		/// Gets the path of the current document.
 		/// </summary>
-		public unsafe string GetFilePath(int bufferId)
+		public unsafe string GetFilePath(IntPtr bufferId)
 		{
 			var path = new StringBuilder(2000);
 			Win32.SendMessage(PluginBase.nppData._nppHandle, (uint) NppMsg.NPPM_GETFULLPATHFROMBUFFERID, bufferId, path);
